Skills Update
Sword/Blunt Weapon Mastery (205)
Dagger Mastery (209)
Mirage (445)
Soul Mastery (467)
Iron Shield (527)
Counter Chance (532)
Counter Rapid Shot (533)
Counter Dash (534)
Counter Mind (535)
Warrior Ability - Haste (651)
Knight Ability - Defense (652)
Rogue Ability - Critical Chance (653)
Wizard Ability - Mana Steal (654)
Enchanter Ability - Barrier (655)
Protection of Rune (755)
Protection of Elemental (756)
Protection of Alignment (757)
Fighter's Will (758)
Archer's Will (759)
Expose Weak Point (767)
Seed of Revenge (761)
Spirit of Phoenix (784)
Eva's Will (786)
Pain of Shilen (788)
Betrayal Mark (792)
Evasion Counter (818)
Patience (983)
Sonic Mastery (992)
Force Mastery (993)
Oblivion (1481)
Weak Constitution (1482)
Thin Skin (1483)
Enervation (1484)
Spite (1485)
Mental Impoverish (1486)
Healer Ability - Heal (1490)
Summoner Ability - Spirit (1491)
Flame Armor (1492)
Frost Armor (1493)
Hurricane Armor (1494)
Excessive Loyalty (1497)
Curse of Life Flow (1511)
Lightning Barrier (1515)
Blessed Blood (1531)
Counter Critical (1542)
Piercing Attack (1564)
Item Skill: Heal (3207)
Item Skill: Blessed Body (3208)
Item Skill: Prayer (3209)
Item Skill: Recharge (3210)
Item Skill: Blessed Soul (3211)
Item Skill: Mana Gain (3212)
Item Skill: Ritual (3213)
Item Skill: Cheer (3214)
Item Skill: Might (3215)
Item Skill: Empower (10)
Item Skill: Duel Might (3217)
Item Skill: Shield (3218)
Item Skill: Magic Barrier (3219)
Item Skill: Duel Weakness (3220)
Item Skill: Agility (3221)
Item Skill: Guidance (3222)
Item Skill: Focus (3223)
Item Skill: Wild Magic (3224)
Item Skill: Charm (3225)
Item Skill: Slow (3226)
Item Skill: Winter (3227)
Item Skill: Stun (3228)
Item Skill: Hold (3229)
Item Skill: Sleep (3230)
Item Skill: Paralyze (3231)
Item Skill: Medusa (3232)
Item Skill: Fear (3233)
Item Skill: Poison (3234)
Item Skill: Bleed (3235)
Item Skill: Silence (3236)
Item Skill: Doom (3237)
Apella Heavy Armor Set (3608)
Apella Light Armor Set (3609)
Apella Robe Set (3610)
PvP Weapon - CP Drain (3650)
PvP Weapon - Cancel (3651)
PvP Weapon - Ignore Shield Defense (3652)
PvP Weapon - Attack Chance (3653)
PvP Weapon - Casting (3654)
PvP Weapon - Rapid Fire (3655)
PvP Weapon - Decrease Range (3656)
PvP Weapon - Decrease Resist (3657)
PvP Shield - Reflect Damage (3658)
PvP Armor - Damage Down (3659)
PvP Armor - Critical Down (3660)
PvP Armor - Speed Down (3662)
PvP Armor - Mirage (3663)
Yellow Talisman of Increase Force (3666)
PvP Armor - Heal (3661)
Sword of Stun (3829)
Rapier of Stun (3834)
Dual-Blade Weapon of Stun (3839)
Dagger of Poison (3861)
Polearm of Stun (3876)
Fist Weapon of Stun (3882)
Blunt Weapon of Stun (3885)
NPC Debuff Shield Slow (4493)
NPC Trigger Stun (5550)
Oblivion Trap (5679)
Critical Hit (5760)
NPC Trigger Bleed (5845)
Burst of Oblivion (5900)
Mist of Oblivion (5901)
Pestilence Cleanse (5911)
Putrefaction Cleanse (5913)
Command of Anchor (5993)
Bless the Blood of Elcadia (6725)
Nightmare Heavy Armor Set - PvP (8193)
Majestic Heavy Armor Set - PvP (8194)
Light Armor of Nightmare Set - PvP (8195)
Light Majestic Leather Armor Set - PvP (8196)
Robe of Nightmare Set - PvP (8197)
Majestic Robe Set - PvP (8198)
Imperial Crusader Breastplate Heavy Armor Set - PvP (8199)
Light Draconic Leather Armor Set - PvP (8200)
Major Arcana Robe Set - PvP (8201)
Dynasty Breastplate - PvP (8202)
Dynasty Breastplate - PvP (8203)
Dynasty Breastplate - PvP (8204)
Dynasty Breastplate - PvP (8205)
Dynasty Breastplate - PvP (8206)
Dynasty Leather Armor - PvP (8207)
Dynasty Leather Armor - PvP (8208)
Dynasty Leather Armor - PvP (8209)
Dynasty Tunic - PvP (8210)
Dynasty Tunic - PvP (8211)
Dynasty Tunic - PvP (8212)
Dynasty Tunic - PvP (8213)
Dynasty Tunic - PvP (8214)
Dynasty Leather Armor - PvP (8215)
Dynasty Leather Armor - PvP (8216)
Dynasty Leather Armor - PvP (8217)
Dynasty Leather Armor - PvP (8218)
Dynasty Platinum Plate - PvP (8219)
Dynasty Platinum Plate - PvP (8220)
Dynasty Platinum Plate - PvP (8221)
Dynasty Platinum Plate - PvP (8222)
Dynasty Jewel Leather Armor - PvP (8223)
Dynasty Jewel Leather Armor - PvP (8224)
Dynasty Jewel Leather Armor - PvP (8225)
Dynasty Jewel Leather Armor - PvP (8226)
Dynasty Jewel Leather Armor - PvP (8227)
Dynasty Jewel Leather Armor - PvP (8228)
Dynasty Silver Satin Tunic - PvP (8229)
Dynasty Silver Satin Tunic - PvP (8230)
Dynasty Silver Satin Tunic - PvP (8231)
Dynasty Silver Satin Tunic - PvP (8232)
Critical Up (8239)
Vesper Heavy Armor Set for Adults (8301)
Vesper Noble Heavy Armor Set for Adults (8302)
Vesper Light Armor Set for Adults (8303)
Vesper Noble Light Armor Set for Adults (8304)
Vesper Robe Set for Adults (8305)
Vesper Noble Robe Set for Adults (8306)
Boing Hammer's Effect of Doubt (8307)
Moirai Heavy Armor Set - PvP (8406)
Moirai Light Armor Set - PvP (8407)
Moirai Robe Set - PvP (8408)
Vorpal Heavy Armor Set - PvP (8409)
Vorpal Leather Armor Set - PvP (8410)
Vorpal Robe Set - PvP (8411)
Elegia Heavy Armor Set - PvP (8412)
Elegia Leather Armor Set - PvP (8413)
Elegia Robe Set - PvP (8414)
Soul Roar (20006)
Mardil (21257)
Giran Designer Cookie (22231)
Giran Designer Cookie (22232)
Master's Blessing - Counter Critical (23246)
Master's Blessing - Counter Critical (23294)

Fixed custom skills:
Mutual Response (1498)
Steal Mana (1526)
Maximum Ability (3426)
Special Ability: Infinity Rapier (3427)
Hellblade (21252)
Butcher Blades (21253)
Claw of Destruction (21254)
Blades of Delusion (21255)
Blood Brother (21256)

New skills:
Shiny Party Hat - Greater Body (21251)

Refactored Trigger skills to use their own target types.
Now each effect uses it's own logger, instead of sharing it across all
effects.

+++ src/main/java/com/l2jserver/datapack/handlers/effecthandlers/pump/TriggerSkillByDamage.java
@@ -43,7 +43,7 @@ public final class TriggerSkillByDamage extends AbstractEffect {
 	private final int _minAttackerLevel;
 	private final int _maxAttackerLevel;
 	private final int _minDamage;
-	private final double _chance;
+	private final int _chance;
 	private final SkillHolder _skill;
 	private final TriggerTargetType _targetType;
 	private final TriggerAttackType _attackerType;
@@ -54,7 +54,7 @@ public final class TriggerSkillByDamage extends AbstractEffect {
 		_minAttackerLevel = params.getInt("minAttackerLevel", 1);
 		_maxAttackerLevel = params.getInt("maxAttackerLevel", 100);
 		_minDamage = params.getInt("minDamage", 1);
-		_chance = params.getDouble("chance", 100);
+		_chance = params.getInt("chance", 100);
 		_skill = new SkillHolder(params.getInt("skillId"), params.getInt("skillLevel", 1));
 		_targetType = params.getEnum("targetType", TriggerTargetType.class, SELF);
 		_attackerType = params.getEnum("attackerType", TriggerAttackType.class, NONE);
